3619: Rename EntityAttributeGrid files
